Module Npy

Author: Laurent Mazare <lmazare@gmail.com>

val write1 : ('a, 'b, 'c) Stdlib.Bigarray.Array1.t -> string -> unit

val write2 : ('a, 'b, 'c) Stdlib.Bigarray.Array2.t -> string -> unit

val write3 : ('a, 'b, 'c) Stdlib.Bigarray.Array3.t -> string -> unit

val write : ?header_len:int -> ('a, 'b, 'c) Stdlib.Bigarray.Genarray.t -> string -> unit

write ?header_len bigarray filename writes a npy file filename with the content of bigarray. header_len can be used to override the npy header length. This is only useful for testing.

module Batch_writer : sig ... end

type packed_array =

| P : (_, _, _) Stdlib.Bigarray.Genarray.t -> packed_array

type packed_array1 =

| P1 : (_, _, _) Stdlib.Bigarray.Array1.t -> packed_array1

type packed_array2 =

| P2 : (_, _, _) Stdlib.Bigarray.Array2.t -> packed_array2

type packed_array3 =

| P3 : (_, _, _) Stdlib.Bigarray.Array3.t -> packed_array3

val read_mmap : string -> shared:bool -> packed_array

read_mmap filename ~shared returns a packed bigarray mmaped to the content of filename. If shared is true modifications made to the array are reflected to the file.

val read_mmap1 : string -> shared:bool -> packed_array1

val read_mmap2 : string -> shared:bool -> packed_array2

val read_mmap3 : string -> shared:bool -> packed_array3

val read_copy : string -> packed_array

val read_copy1 : string -> packed_array1

val read_copy2 : string -> packed_array2

val read_copy3 : string -> packed_array3

module Npz : sig ... end

Conversion functions from packed arrays to bigarrays

val to_bigarray : 'c Stdlib.Bigarray.layout -> ('a, 'b) Stdlib.Bigarray.kind -> packed_array -> ('a, 'b, 'c) Stdlib.Bigarray.Genarray.t option

to_bigarray layout kind packed_array returns Some a with a a Bigarray.Genarray.t if the layout and the kind of packed_array were equal to the layout and kind arguments. Otherwise, to_bigarray returns None

val to_bigarray1 : 'c Stdlib.Bigarray.layout -> ('a, 'b) Stdlib.Bigarray.kind -> packed_array1 -> ('a, 'b, 'c) Stdlib.Bigarray.Array1.t option

Same as to_bigarray for Bigarray.Array1.t

val to_bigarray2 : 'c Stdlib.Bigarray.layout -> ('a, 'b) Stdlib.Bigarray.kind -> packed_array2 -> ('a, 'b, 'c) Stdlib.Bigarray.Array2.t option

Same as to_bigarray for Bigarray.Array2.t

val to_bigarray3 : 'c Stdlib.Bigarray.layout -> ('a, 'b) Stdlib.Bigarray.kind -> packed_array3 -> ('a, 'b, 'c) Stdlib.Bigarray.Array3.t option

Same as to_bigarray for Bigarray.Array3.t